implemented steps
